pyproject.toml
[project]
name = "image-tools"
version = "0.1.0"
description = "Tools for image processing and conversion, including HTML presentation to image converter"
readme = "README.md"
requires-python = ">=3.13"
dependencies = [
    "playwright>=1.40.0",
    "beautifulsoup4>=4.12.0",
    "pillow>=10.0.0",
]
